# swagger-documentation.yaml
openapi: 3.0.0
info:
  title: Protobin API
  version: 1.0.0
  description: |
    ### Общая документация API для всех сервисов Protobin
    
    Это модуль для работы с проектами, здесь вы можете получить:
    - теги
    - стек
    - проекты
    - участников проекта
    - права участников
    - фрагменты
    - элементы доски
    - пользователей
    - авторизацию
    и т.д.
  contact:
    name: API Support
    url: https://protobin.com
    email: vogistv@gmail.com
  license:
    name: Proprietary
    url: https://protobin.com
  termsOfService: https://protobin.com
servers:
  - url: http://localhost:8080
    description: Локальная среда
  - url: https://api.protobin.com
    description: Продакшн
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Authorization header using the Bearer scheme
  schemas:
    ResponseDto:
      type: object
      title: Ответ
      description: Стандартизированный ответ от API
      properties:
        status:
          type: integer
          title: Статус
          description: Статус ответа
          example: 200
        message:
          type: string
          title: Сообщение
          description: Сообщение ответа
          example: "Какой-либо ответ"
...
    StackTechnologyResponseDto:
      type: object
      title: Технология стэка
      description: Поля получаемые с технологии стэка
      properties:
        id:
          type: string
          format: uuid
          description: Идентификатор технологии
          example: "3d5f1a26-c554-4bfe-811a-0e9ef2ed52fc"
        name:
          type: string
          description: Название технологии
          example: "java"
...
  responses:
    BadRequest:
      description: Неправильное тело запроса
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResponseDto'
    Unauthorized:
      description: Не авторизован
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResponseDto'
    Forbidden:
      description: Недостаточно прав
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResponseDto'
    NotFound:
      description: Ресурс был не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResponseDto'
paths:
...
  /v1/stack:
    post:
      tags:
        - Stack Technologies
      summary: Создать список технологий стэка
      description: Создается список технологий для указанного проекта
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StackTechnologyCreateDto'
      responses:
        '201':
          description: Тэги созданы
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StackTechnologyResponseDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/stack/{id}:
    delete:
      tags:
        - Stack Technologies
      summary: Удалить технологию стэка
      description: Удаляется технология стэка
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Идентификатор технологии стэка
      responses:
        '204':
          description: Была удалена технология стэка
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...